backend/requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
openpyxl
pandas
rest_framework_simplejwt
